moved epithelial cell creation to its own file
