+++ client/components/GameFunctions/createFunctions/index.js
@@ -2,5 +2,6 @@ import { players } from './addPlayers'
 import { keyboardControls } from './keyboardControls'
 import { scoreAndStars } from './lifeStars';
 import { NPCCells } from './cells'
+import { epithelialCells, makeEpithelialCell, epithelialCellCollision } from './epithelialCells'
 
-export {players, keyboardControls, scoreAndStars, NPCCells }
+export {players, keyboardControls, scoreAndStars, NPCCells, epithelialCells, makeEpithelialCell, epithelialCellCollision }
