+++ client/components/GameFunctions/update.js
@@ -119,13 +119,15 @@ export function update(time) {
   }
 
   this.antibodies.getChildren().forEach(antibody => {
-    this.badGuys.forEach(badGuy => 
+    console.log(antibody)
+    this.badGuys.forEach(badGuy => {
+      console.log(badGuy)
       overlapCollision.call(this, {x: antibody.x, y: antibody.y}, badGuy, () => {
         console.log('owie!')
         // antibody.setActive(false)
         // antibody.setVisible(false)
         antibody.destroy()
       })
-    )
+    })
   })
 }
